+++ index.html
@@ -2,7 +2,6 @@
 <html>
   <head>
     <script defer type="text/javascript" src="dist/reactions.js"></script>
-     <link rel="stylesheet" type="text/css" href="src/main.css">
   </head>
   <body></body>
 </html>
